Datatype change for HistorianStep in DeviceObservations to integer

+++ src/Masterloop.Core.Types/Devices/DeviceObservation.cs
@@ -14,7 +14,7 @@ namespace Masterloop.Core.Types.Devices
 
         public HistorianType Historian { get; set; }
 
-        public double HistorianStep { get; set; }
+        public int HistorianStep { get; set; }
 
         public int Quantity { get; set; }
